backend/requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
pymongo==4.16.0
pydantic==2.12.5
python-dotenv==1.2.1
email-validator==2.3.0
